[core] Refactor move Internet to feature_utils

[tests] Fix internet tests
