+++ indexer/feature_utils.hpp
@@ -16,6 +16,20 @@ namespace feature
   static constexpr std::string_view kAtmSymbol = "💳";
   static constexpr std::string_view kWheelchairSymbol = "♿️";
 
+  /// OSM internet_access tag values.
+  enum class Internet
+  {
+    Unknown,  //!< Internet state is unknown (default).
+    Wlan,     //!< Wireless Internet access is present.
+    Terminal, //!< A computer with internet service.
+    Wired,    //!< Wired Internet access is present.
+    Yes,      //!< Unspecified Internet access is available.
+    No        //!< There is definitely no any Internet access.
+  };
+  std::string DebugPrint(Internet internet);
+  /// @param[in]  inet  Should be lowercase like in DebugPrint.
+  Internet InternetFromString(std::string_view inet);
+
   // Address house numbers interpolation.
   enum class InterpolType { None, Odd, Even, Any };
